archive/refactor_prototype/requirements.txt
qtpy
Pillow
qt-material